+++ lib/db_agent/webapp.rb
@@ -1,10 +1,17 @@
 module DbAgent
   class Webapp < Sinatra::Base
 
-    set :raise_errors, true
+    set :raise_errors, false
     set :show_exceptions, false
     set :dump_errors, true
     set :db_handler, DbAgent.default_handler
+
+    error do
+      route = env['sinatra.route']
+      params = env['sinatra.error.params']
+      error = env['sinatra.error']
+      "DbAgent (route: #{route} | params: #{params}) error: #{error.message}\n#{error.backtrace[0..10]}\n\n"
+    end
 
     get '/ping' do
       settings.db_handler.sequel_db.test_connection